Add files via upload
This script can be used in QuPath to detect all stages of Plasmodium falciparum stained with DAPI. Gametocyte detection is not perfect in this script. For asexual stages together, sensitivity is 99.6 % and specificity is 96.25 % (developed to use as denominator of sexual conversion rate calculations when added to gametocyte detection), do not use to try to sub classify classify asexual life stages. Has to be used with classifier "250604_classfier_AllStagesB"